move AsyncAPI objects to a separate package
