+++ ogcapi-draft/ogcapi-pubsub/src/main/java/de/ii/ogcapi/pubsub/app/AsyncApiJson.java
@@ -15,8 +15,8 @@ import de.ii.ogcapi.foundation.domain.ApiRequestContext;
 import de.ii.ogcapi.foundation.domain.ExtensionConfiguration;
 import de.ii.ogcapi.foundation.domain.ImmutableApiMediaType;
 import de.ii.ogcapi.foundation.domain.ImmutableApiMediaTypeContent;
-import de.ii.ogcapi.pubsub.domain.AsyncApi;
 import de.ii.ogcapi.pubsub.domain.AsyncApiDefinitionFormatExtension;
+import de.ii.ogcapi.pubsub.domain.asyncapi.AsyncApi;
 import io.swagger.v3.oas.models.media.ObjectSchema;
 import java.util.Optional;
 import javax.inject.Inject;
