+++ ogcapi-draft/ogcapi-pubsub/src/main/java/de/ii/ogcapi/pubsub/app/QueriesHandlerPubSubImpl.java
@@ -19,25 +19,25 @@ import de.ii.ogcapi.foundation.domain.OgcApiDataV2;
 import de.ii.ogcapi.foundation.domain.QueryHandler;
 import de.ii.ogcapi.foundation.domain.QueryInput;
 import de.ii.ogcapi.foundation.domain.QueryInputGeneric;
-import de.ii.ogcapi.pubsub.domain.AsyncApi;
-import de.ii.ogcapi.pubsub.domain.AsyncApiChannel;
 import de.ii.ogcapi.pubsub.domain.AsyncApiDefinitionFormatExtension;
-import de.ii.ogcapi.pubsub.domain.AsyncApiReference;
-import de.ii.ogcapi.pubsub.domain.AsyncApiServer;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApi;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiChannel;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiComponents;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiContact;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiInfo;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiLicense;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiMessage;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiOperation;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiOperationBindingsMqtt;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiReference;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiSecurity;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiServer;
-import de.ii.ogcapi.pubsub.domain.ImmutableAsyncApiServerBindingsMqtt;
 import de.ii.ogcapi.pubsub.domain.QueriesHandlerPubSub;
+import de.ii.ogcapi.pubsub.domain.asyncapi.AsyncApi;
+import de.ii.ogcapi.pubsub.domain.asyncapi.AsyncApiChannel;
+import de.ii.ogcapi.pubsub.domain.asyncapi.AsyncApiReference;
+import de.ii.ogcapi.pubsub.domain.asyncapi.AsyncApiServer;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApi;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiChannel;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiComponents;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiContact;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiInfo;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiLicense;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiMessage;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiOperation;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiOperationBindingsMqtt;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiReference;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiSecurity;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiServer;
+import de.ii.ogcapi.pubsub.domain.asyncapi.ImmutableAsyncApiServerBindingsMqtt;
 import java.net.URI;
 import java.text.MessageFormat;
 import java.util.Date;
