+++ ogcapi-draft/ogcapi-pubsub/src/main/java/de/ii/ogcapi/pubsub/domain/asyncapi/AsyncApi.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.ogcapi.pubsub.domain;
+package de.ii.ogcapi.pubsub.domain.asyncapi;
 
 import java.util.Map;
 import org.immutables.value.Value;
