+++ ogcapi-draft/ogcapi-pubsub/src/main/java/de/ii/ogcapi/pubsub/domain/asyncapi/AsyncApiChannel.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.ogcapi.pubsub.domain;
+package de.ii.ogcapi.pubsub.domain.asyncapi;
 
 import java.util.List;
 import java.util.Map;
@@ -16,6 +16,7 @@ import org.immutables.value.Value;
 public interface AsyncApiChannel {
 
   Optional<String> getDescription();
+
   // Not implemented
   // AsyncApiOperation getPublish();
   AsyncApiOperation getSubscribe();
