+++ ogcapi-draft/ogcapi-pubsub/src/main/java/de/ii/ogcapi/pubsub/domain/asyncapi/AsyncApiChannelBindingsMqtt.java
@@ -5,13 +5,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.ogcapi.pubsub.domain;
+package de.ii.ogcapi.pubsub.domain.asyncapi;
 
-import de.ii.ogcapi.features.core.domain.JsonSchema;
 import org.immutables.value.Value;
 
 @Value.Immutable
-public interface AsyncApiParameter {
-
-  JsonSchema getSchema();
-}
+public interface AsyncApiChannelBindingsMqtt {}
