+++ ogcapi-draft/ogcapi-pubsub/src/main/java/de/ii/ogcapi/pubsub/domain/asyncapi/AsyncApiComponents.java
@@ -5,14 +5,18 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.ogcapi.pubsub.domain;
+package de.ii.ogcapi.pubsub.domain.asyncapi;
 
+import de.ii.ogcapi.features.core.domain.JsonSchema;
+import java.util.Map;
 import org.immutables.value.Value;
 
 @Value.Immutable
-public interface AsyncApiMessageBindingsMqtt {
+public interface AsyncApiComponents {
 
-  default String getBindingVersion() {
-    return "0.1.0";
-  }
+  Map<String, AsyncApiMessage> getMessages();
+
+  Map<String, JsonSchema> getSchemas();
+
+  Map<String, AsyncApiSecurity> getSecuritySchemes();
 }
