+++ ogcapi-draft/ogcapi-pubsub/src/main/java/de/ii/ogcapi/pubsub/domain/asyncapi/AsyncApiMessage.java
@@ -5,25 +5,29 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.ogcapi.pubsub.domain;
+package de.ii.ogcapi.pubsub.domain.asyncapi;
 
 import de.ii.ogcapi.features.core.domain.JsonSchema;
-import java.util.Map;
 import java.util.Optional;
 import org.immutables.value.Value;
 
 @Value.Immutable
-public interface AsyncApiOperation {
+public interface AsyncApiMessage {
 
-  Optional<String> getOperationId();
+  Optional<String> getName();
+
+  Optional<String> getTitle();
 
   Optional<String> getSummary();
 
   Optional<String> getDescription();
 
-  Optional<AsyncApiOperationBindingsMqtt> getBindings();
+  @Value.Default
+  default String getContentType() {
+    return "application/geo+json";
+  }
 
-  Map<String, JsonSchema> getParameters();
+  Optional<JsonSchema> getPayload();
 
-  AsyncApiReference getMessage();
+  Optional<AsyncApiMessageBindingsMqtt> getBindings();
 }
