+++ ogcapi-draft/ogcapi-pubsub/src/main/java/de/ii/ogcapi/pubsub/domain/asyncapi/AsyncApiOperation.java
@@ -5,18 +5,25 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.ogcapi.pubsub.domain;
+package de.ii.ogcapi.pubsub.domain.asyncapi;
 
 import de.ii.ogcapi.features.core.domain.JsonSchema;
 import java.util.Map;
+import java.util.Optional;
 import org.immutables.value.Value;
 
 @Value.Immutable
-public interface AsyncApiComponents {
+public interface AsyncApiOperation {
 
-  Map<String, AsyncApiMessage> getMessages();
+  Optional<String> getOperationId();
 
-  Map<String, JsonSchema> getSchemas();
+  Optional<String> getSummary();
 
-  Map<String, AsyncApiSecurity> getSecuritySchemes();
+  Optional<String> getDescription();
+
+  Optional<AsyncApiOperationBindingsMqtt> getBindings();
+
+  Map<String, JsonSchema> getParameters();
+
+  AsyncApiReference getMessage();
 }
