+++ ogcapi-draft/ogcapi-pubsub/src/main/java/de/ii/ogcapi/pubsub/domain/asyncapi/AsyncApiReference.java
@@ -5,14 +5,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.ogcapi.pubsub.domain;
+package de.ii.ogcapi.pubsub.domain.asyncapi;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.Optional;
 import org.immutables.value.Value;
 
 @Value.Immutable
-public interface AsyncApiSecurity {
+public interface AsyncApiReference {
 
-  default String getType() {
-    return "userPassword";
-  }
+  @JsonProperty("$ref")
+  Optional<String> getRef();
 }
