+++ ogcapi-draft/ogcapi-pubsub/src/main/java/de/ii/ogcapi/pubsub/domain/asyncapi/AsyncApiSecurity.java
@@ -5,24 +5,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.ogcapi.pubsub.domain;
+package de.ii.ogcapi.pubsub.domain.asyncapi;
 
-import java.util.Optional;
 import org.immutables.value.Value;
 
 @Value.Immutable
-public interface AsyncApiInfo {
+public interface AsyncApiSecurity {
 
-  String getTitle();
-
-  Optional<String> getDescription();
-
-  @Value.Default
-  default String getVersion() {
-    return "1.0.0";
+  default String getType() {
+    return "userPassword";
   }
-
-  Optional<AsyncApiContact> getContact();
-
-  Optional<AsyncApiLicense> getLicense();
 }
